+++ src/file.hpp
@@ -49,8 +49,8 @@ public:
   }
 
   void open_read(const std::function<void(FileStream*)> &cb);
-  void open_read(int seek, const std::function<void(FileStream*)> &cb);
-  void open_read(int seek, int size, const std::function<void(FileStream*)> &cb);
+  void open_read(size_t seek, const std::function<void(FileStream*)> &cb);
+  void open_read(size_t seek, size_t size, const std::function<void(FileStream*)> &cb);
   void open_write(bool append);
   void write(const Data &data);
   void close();
