+++ src/filters/read.hpp
@@ -40,8 +40,8 @@ class File;
 class Read : public Filter, public EventSource {
 public:
   struct Options : public pipy::Options {
-    int seek = 0;
-    int size = -1;
+    size_t seek = 0;
+    size_t size = -1;
     pjs::Ref<pjs::Function> seek_f;
     pjs::Ref<pjs::Function> size_f;
     Options() {}
